chore(application-system): prevent errors on prune application charge (#20301)
* added some checks to prevent or catch errors

* chore: format update dirty file

* moved 404 check to chargeFjsV2Client.service.ts to match previous functionality

* chore: format update dirty file

---------

+++ apps/application-system/api/src/app/modules/application/charge/application-charge.service.ts
@@ -55,13 +55,16 @@ export class ApplicationChargeService {
       const paymentUrl = JSON.parse(payment.definition as unknown as string)
         .paymentUrl as string
 
-      const url = new URL(paymentUrl)
-      const chargeId = url.pathname.split('/').pop()
+      //paymentUrl will be undefined if using MockPayment on dev causing an error
+      if (paymentUrl) {
+        const url = new URL(paymentUrl)
+        const chargeId = url.pathname.split('/').pop()
 
-      this.logger.info('deleteCharge chargeId', chargeId)
+        this.logger.info('deleteCharge chargeId', chargeId)
 
-      if (chargeId) {
-        await this.chargeFjsV2ClientService.deleteCharge(chargeId)
+        if (chargeId) {
+          await this.chargeFjsV2ClientService.deleteCharge(chargeId)
+        }
       }
     } catch (error) {
       this.logger.error(
